+++ R/vbFuns.R
@@ -8,7 +8,7 @@
 #' 
 #' @return A function that can be used to predict fish length given a vector of ages and values for the model parameters and, in some parameterizations, values for some constants.  The result should be saved to an object that can then be used as a function name.  When the resulting function is used the parameters are ordered as shown when the definitions of the parameters are printed after the function is called (if \code{msg=TRUE}).
 #'
-#'If \code{simple=FALSE} then the values for all parameters can be included as a vector in the first parameter argument.  Similarly, the values for all constants can be included as a vector in the first constant argument (i.e., \code{t1}).  If \code{simple=TRUE} then all parameters and constants must be declared individually.  If \code{simple=TRUE} then the resulting function is somewhat easier to read.
+#'If \code{simple=FALSE} then the values for all parameters can be included as a vector in the first parameter argument.  Similarly, the values for all constants can be included as a vector in the first constant argument (i.e., \code{t1}).  If \code{simple=TRUE} then all parameters and constants must be declared individually.  The resulting function is somewhat easier to read when \code{simple=TRUE}.
 #'
 #' @author Derek H. Ogle, \email{dogle@@northland.edu}, thanks to Gabor Grothendieck for a hint about using \code{get()}.
 #'